Sprite.Background finished all backgrounds
Background.png is the static background

Midground.png is a parallaxing midground that loops into itself

Foreground.png goes in front of everything else, to be put on the bottom
of the screen and overlap in front of the Midground/Background. It
should move with the stage and should be topped with platform toppers
when applicable.

Block.png are the blocks that should make up any platforming. They
should move with the Foreground and be topped with platform toppers when
applicable.

Platform Topper.png goes on top of the Foreground or Blocks, these are
what the main character and enemies are capable of walking/running on.

Platform Topper_alt.png is a differant design of platform Topper.